Combine WOfS code into one file.
It's small enough to be more readable when together.
