+++ src/wofs_ingester/__init__.py
@@ -32,6 +32,7 @@ Ingestion of water tiles.
 
 import logging
 from ..abstract_ingester.pretiled import GdalMdDataset, PreTiledIngester
+import sys
 
 
 _LOG = logging.getLogger(__name__)
@@ -91,3 +92,24 @@ class WofsDataset(GdalMdDataset):
     """
     pass
 
+
+def run_ingest():
+    """
+
+    :return:
+    """
+    logging.basicConfig(stream=sys.stdout,
+                        format='%(asctime)s %(name)s %(levelname)s %(message)s',
+                        level=logging.INFO)
+
+    ingester = WofsIngester()
+
+    if ingester.args.debug:
+        logging.getLogger().setLevel(logging.DEBUG)
+
+    ingester.ingest(ingester.args.source_dir)
+
+    ingester.collection.cleanup()
+
+if __name__ == '__main__':
+    run_ingest()
